Move random_gif_path fixture to conftest.py
Previously I defined the fixture in test_collection.py and used imports
in the other two modules. Unfortunately, ruff wasn't very happy with
that because it considered them as unused imports (F401) and the
parameter with the same name as a redefinition (F811).
